chore: 更新 controls/core/graph/interactive 等模块与组件文档生成脚本

+++ src/core/InstancedMesh2/InstancedEntity.ts
@@ -180,7 +180,7 @@ export class InstancedEntity {
     const id = this.id;
     const offset = id * MATRIX_ELEMENTS;
 
-    writeMatrixFromTransform(te as Float32Array, offset, quaternion, scale, position);
+    writeMatrixFromTransform(te, offset, quaternion, scale, position);
 
     owner.matricesTexture.enqueueUpdate(id);
